+++ app/test_helpers.go
@@ -57,7 +57,6 @@ var DefaultConsensusParams = &abci.ConsensusParams{
 }
 
 func setup(withGenesis bool, invCheckPeriod uint) (*App, GenesisState) {
-	SetConfig()
 	db := dbm.NewMemDB()
 	encCdc := MakeEncodingConfig()
 	app := New(log.NewNopLogger(), db, nil, true, map[int64]bool{}, DefaultNodeHome, invCheckPeriod, encCdc, EmptyAppOptions{})
