+++ x/cronos/handler_test.go
@@ -8,111 +8,127 @@ import (
 	"github.com/crypto-org-chain/cronos/app"
 	"github.com/crypto-org-chain/cronos/x/cronos"
 	"github.com/crypto-org-chain/cronos/x/cronos/types"
-	"github.com/stretchr/testify/require"
+	"github.com/stretchr/testify/suite"
 	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
+	"github.com/tharsis/ethermint/crypto/ethsecp256k1"
 	"strings"
 	"testing"
+	"time"
 )
 
-func TestInvalidMsg(t *testing.T) {
-	app := app.Setup(false)
-	handler := cronos.NewHandler(app.CronosKeeper)
+type CronosTestSuite struct {
+	suite.Suite
 
-	res, err := handler(sdk.NewContext(nil, tmproto.Header{}, false, nil), testdata.NewTestMsg())
-	require.Error(t, err)
-	require.Nil(t, res)
-
-	_, _, log := sdkerrors.ABCIInfo(err, false)
-	require.True(t, strings.Contains(log, "unrecognized cronos message type"))
+	ctx     sdk.Context
+	handler sdk.Handler
+	app     *app.App
+	address sdk.AccAddress
 }
 
-func TestMsgConvertVouchers(t *testing.T) {
+func TestCronosTestSuite(t *testing.T) {
+	suite.Run(t, new(CronosTestSuite))
+}
+
+func (suite *CronosTestSuite) SetupTest() {
+	checkTx := false
+	suite.app = app.Setup(false)
+	suite.ctx = suite.app.BaseApp.NewContext(checkTx, tmproto.Header{Height: 1, ChainID: app.TestAppChainID, Time: time.Now().UTC()})
+	suite.handler = cronos.NewHandler(suite.app.CronosKeeper)
+
+	privKey, err := ethsecp256k1.GenerateKey()
+	suite.Require().NoError(err)
+	suite.address = sdk.AccAddress(privKey.PubKey().Address())
+}
+
+func (suite *CronosTestSuite) TestInvalidMsg() {
+	res, err := suite.handler(sdk.NewContext(nil, tmproto.Header{}, false, nil), testdata.NewTestMsg())
+	suite.Require().Error(err)
+	suite.Nil(res)
+
+	_, _, log := sdkerrors.ABCIInfo(err, false)
+	suite.Require().True(strings.Contains(log, "unrecognized cronos message type"))
+}
+
+func (suite *CronosTestSuite) TestMsgConvertVouchers() {
 	testCases := []struct {
-		name     string
-		msg   *types.MsgConvertVouchers
-		malleate   func()
+		name          string
+		msg           *types.MsgConvertVouchers
+		malleate      func()
 		expectedError error
 	}{
 		{
 			"Wrong address",
 			types.NewMsgConvertVouchers("test", sdk.NewCoins(sdk.NewCoin("aphoton", sdk.NewInt(1)))),
-			func(){},
+			func() {},
 			errors.New("decoding bech32 failed: invalid bech32 string length 4"),
 		},
 		{
 			"Empty address",
 			types.NewMsgConvertVouchers("", sdk.NewCoins(sdk.NewCoin("aphoton", sdk.NewInt(1)))),
-			func(){},
+			func() {},
 			errors.New("empty address string is not allowed"),
 		},
 		{
 			"Correct address with non supported coin denom",
-			types.NewMsgConvertVouchers("eth1cml96vmptgw99syqrrz8az79xer2pcgpl6nuct", sdk.NewCoins(sdk.NewCoin("fake", sdk.NewInt(1)))),
-			func(){},
+			types.NewMsgConvertVouchers(suite.address.String(), sdk.NewCoins(sdk.NewCoin("fake", sdk.NewInt(1)))),
+			func() {},
 			nil,
 		},
 	}
 
-	app := app.Setup(false)
-	ctx := app.BaseApp.NewContext(false, tmproto.Header{})
-
 	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			handler := cronos.NewHandler(app.CronosKeeper)
-			_, err := handler(ctx, tc.msg)
+		suite.Run(tc.name, func() {
+			handler := cronos.NewHandler(suite.app.CronosKeeper)
+			_, err := handler(suite.ctx, tc.msg)
 			if tc.expectedError != nil {
-				require.EqualError(t, err, tc.expectedError.Error())
+				suite.Require().EqualError(err, tc.expectedError.Error())
 			} else {
-				require.NoError(t, err)
+				suite.Require().NoError(err)
 			}
 		})
 	}
 }
 
-func TestMsgTransferTokens(t *testing.T) {
+func (suite *CronosTestSuite) TestMsgTransferTokens() {
 	testCases := []struct {
-		name     string
-		msg   *types.MsgTransferTokens
-		malleate   func()
+		name          string
+		msg           *types.MsgTransferTokens
+		malleate      func()
 		expectedError error
 	}{
 		{
 			"Wrong from address",
 			types.NewMsgTransferTokens("test", "to", sdk.NewCoins(sdk.NewCoin("aphoton", sdk.NewInt(1)))),
-			func(){},
+			func() {},
 			errors.New("decoding bech32 failed: invalid bech32 string length 4"),
 		},
 		{
 			"Empty from address",
 			types.NewMsgTransferTokens("", "to", sdk.NewCoins(sdk.NewCoin("aphoton", sdk.NewInt(1)))),
-			func(){},
+			func() {},
 			errors.New("empty address string is not allowed"),
 		},
 		{
 			"Empty to address",
-			types.NewMsgTransferTokens("eth1cml96vmptgw99syqrrz8az79xer2pcgpl6nuct", "", sdk.NewCoins(sdk.NewCoin("aphoton", sdk.NewInt(1)))),
-			func(){},
+			types.NewMsgTransferTokens(suite.address.String(), "", sdk.NewCoins(sdk.NewCoin("aphoton", sdk.NewInt(1)))),
+			func() {},
 			errors.New("to address cannot be empty"),
 		},
 		{
 			"Correct address with non supported coin denom",
-			types.NewMsgTransferTokens("eth1cml96vmptgw99syqrrz8az79xer2pcgpl6nuct", "to", sdk.NewCoins(sdk.NewCoin("fake", sdk.NewInt(1)))),
-			func(){},
+			types.NewMsgTransferTokens(suite.address.String(), "to", sdk.NewCoins(sdk.NewCoin("fake", sdk.NewInt(1)))),
+			func() {},
 			nil,
 		},
 	}
-
-	app := app.Setup(false)
-	ctx := app.BaseApp.NewContext(false, tmproto.Header{})
-
 	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			handler := cronos.NewHandler(app.CronosKeeper)
-			_, err := handler(ctx, tc.msg)
+		suite.Run(tc.name, func() {
+			handler := cronos.NewHandler(suite.app.CronosKeeper)
+			_, err := handler(suite.ctx, tc.msg)
 			if tc.expectedError != nil {
-				require.EqualError(t, err, tc.expectedError.Error())
+				suite.Require().EqualError(err, tc.expectedError.Error())
 			} else {
-				require.NoError(t, err)
+				suite.Require().NoError(err)
 			}
 		})
 	}
